Skip MIME Type test if MIME_Type is not available.
It only ever will be available for someone who has MIME_Type
installed in such a way that it will be autoloaded correctly.
I'm not sure that is even possible so this may make this test
a noop.

MIME_Type never will be available on Travis without work and
this Detector will likely be phased out before that work can
be done so this test will always be skipped on Travis.